Refactor filter functionality into own custom hook
