+++ src/containers/ANC/list/index.tsx
@@ -1,39 +1,9 @@
 /** Presentational component and container for the ANC listing page */
-import React, { useState } from 'react';
-
-/** types for a function used as to filter */
-type FilterFunction = <T>(dataObj: T) => boolean;
-
-/** describes how filter state information will be stored */
-interface FilterState {
-  [k: string]: {
-    filter: FilterFunction;
-    [key: string]: any;
-  };
-}
-
-/** interface for dataObj handled by setFilterState */
-interface FilterStateAction {
-  filterId: string;
-  filterFunction: FilterFunction;
-  overwrite: boolean;
-  properties: { [key: string]: any };
-}
+import React from 'react';
+import { useFilters } from './hooks';
 
 const ANCListView: React.FC<{}> = () => {
-  const [filterState, setFilterState] = useState<FilterState>({});
-
-  const addFilter = (action: FilterStateAction) => {
-    const filterToAdd: FilterState = action.overwrite
-      ? {
-          [action.filterId]: { filter: action.filterFunction, ...action.properties },
-        }
-      : {
-          ...filterState,
-          [action.filterId]: { filter: action.filterFunction, ...action.properties },
-        };
-    setFilterState(filterToAdd);
-  };
+  const [filterState, addFilter, setFilterState] = useFilters();
 
   return <div />;
 };
